requirements.txt
ifaddr>=0.1.0
jsonschema>=3.2.0
pyee>=7.0.0
zeroconf>=0.27.0
ujson
pyjwt
httpx==0.14.*
starlette==0.13.*
uvicorn==0.11.*
tortoise-orm==0.16.*
